requirements.txt
# Kalshi Prediction Bot - Python Dependencies
# Install with: pip install -r requirements.txt

# Core async framework
aiohttp>=3.9.0
websockets>=12.0
aiosqlite>=0.19.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Machine Learning (for paper trading ML integration)
scikit-learn>=1.3.0

# Cryptography for Kalshi API signing
cryptography>=41.0.0

# Configuration
python-dotenv>=1.0.0

# HTTP client (sync fallback)
requests>=2.31.0

# Date/time handling
python-dateutil>=2.8.0

# Rate limiting
aiolimiter>=1.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type checking
mypy>=1.5.0

# Logging
structlog>=23.0.0

# Additional utilities
tenacity>=8.2.0  # Retry logic
cachetools>=5.3.0  # Caching
